+++ src/utilities.js
@@ -27,18 +27,18 @@ export function deepCopy (target) {
 
 export function extend (destination, ...args) {
   args.forEach(source => {
-    if (!source) return
-
-    Object.keys(source).forEach(property => {
-      if (source[property] && isPlainObject(source[property])) {
-        if (!hasOwnProperty(destination, property)) destination[property] = {}
-        extend(destination[property], source[property])
-      } else if (Array.isArray(source[property])) {
-        destination[property] = deepCopy(source[property])
-      } else {
-        destination[property] = source[property]
-      }
-    })
+    if (source) {
+      Object.keys(source).forEach(property => {
+        if (source[property] && isPlainObject(source[property])) {
+          if (!hasOwnProperty(destination, property)) destination[property] = {}
+          extend(destination[property], source[property])
+        } else if (Array.isArray(source[property])) {
+          destination[property] = deepCopy(source[property])
+        } else {
+          destination[property] = source[property]
+        }
+      })
+    }
   })
 
   return destination