+++ CHANGELOG.md
@@ -1,11 +1,29 @@
 ### Unreleased
+
 - added option `max_depth` used to specify the maximum depth of level's schema that have to be rendered
 - added option `use_default_values` used to specify if default values based on the "type" of the property have to be used
 - when `use_default_values` is false, number and integer fields have undefined value when input is empty
 - when `use_default_values` is false, string fields have undefined value at the beginning. After that user edit the field, empty input is threated as an empty string
 - added option `validation_error_placement` to change position of errors logs under each field 
 
-### 2.3.0-dev
+### 2.5.1
+
+- Fix for #837 and chekboxes not displaying error messages #843
+
+### 2.5.0
+
+- added new stepper editor (with default HTML & Bootstrap 4 Theme)
+- fixed babel transpiling bug
+- updated docs
+
+### 2.4.0
+
+- Fixed using multiple dependencies for an editor. Fix #703
+- Evaluate all dependencies for an editor to determine visibility
+- Fixed meta-schema usage
+- Added `min`, `max` for `dateTime`
+
+### 2.3.0
 
 - Removed codeception container, use `codeceptjs` as node-dev module
 - Enabled Bootstrap 3 theme/icons